Spring added

+++ src/main/java/by/clevertec/entity/Car.java
@@ -15,13 +15,13 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-
-import java.util.ArrayList;
-import java.util.List;
-
 import lombok.ToString;
 import org.hibernate.annotations.Cache;
 import org.hibernate.annotations.CacheConcurrencyStrategy;
+
+import java.math.BigDecimal;
+import java.util.ArrayList;
+import java.util.List;
 
 @Entity
 @Table(name = "cars")
@@ -46,7 +46,7 @@ public class Car {
     private int year;
 
     @Column(nullable = false)
-    private double price;
+    private BigDecimal price;
 
     @ToString.Exclude
     @ManyToOne
